requirements.txt
fastapi==0.110.0 
uvicorn==0.27.1 

#new
SQLAlchemy==2.0.28
alembic==1.13.1

python-dotenv==1.0.1
python-jose==3.3.0

requests==2.31.0